+++ equalsverifier-core/src/test/java/nl/jqno/equalsverifier/integration/extra_features/JpaIdTest.java
@@ -410,6 +410,11 @@ public class JpaIdTest {
             .assertMessageContains("Significant fields");
     }
 
+    @Test
+    public void succeed_whenEqualsIsImplementedInSuperclass_givenWarningSurrogateKeyIsSuppressed() {
+        EqualsVerifier.forClass(SubclassEntity.class).suppress(Warning.SURROGATE_KEY).verify();
+    }
+
     static class JpaIdBusinessKeyPerson {
 
         @Id
@@ -1047,6 +1052,40 @@ public class JpaIdTest {
         @Override
         public final int hashCode() {
             return Objects.hash(socialSecurity);
+        }
+    }
+
+    static class AbstractEntity {
+
+        @Id
+        private final UUID id;
+
+        public AbstractEntity(UUID id) {
+            this.id = id;
+        }
+
+        @Override
+        public final boolean equals(Object obj) {
+            if (!(obj instanceof AbstractEntity)) {
+                return false;
+            }
+            AbstractEntity other = (AbstractEntity) obj;
+            return Objects.equals(id, other.id);
+        }
+
+        @Override
+        public final int hashCode() {
+            return Objects.hash(id);
+        }
+    }
+
+    static class SubclassEntity extends AbstractEntity {
+
+        private final String name;
+
+        public SubclassEntity(UUID id, String name) {
+            super(id);
+            this.name = name;
         }
     }
 }